[chore] Remove orders.py
