+++ app/models/__init__.py
@@ -1,7 +1,7 @@
 """This file contains the database class that contains all database related methods"""
 from pprint import pprint
 from urllib.parse import urlparse
-
+from flask import jsonify
 import psycopg2
 import os
 
@@ -106,3 +106,49 @@ class Database:
             "SELECT role FROM users WHERE email = '%s'" % (email))
         item = self.cursor.fetchone()
         return item
+
+    def add_order(self, order):
+        """add a new order """
+        insert_command = "INSERT INTO orders(menu_id, user_id, quantity, status, created_at) Values (%s, %s, %s, '%s','%s');" % (
+            order.menu_id, order.user_id, order.quantity, order.status, order.created_at)
+        try:
+            self.cursor.execute(insert_command)
+            self.cursor.execute(
+                "SELECT row_to_json(row) FROM (SELECT * FROM orders WHERE user_id = %s AND created_at = '%s') row;" % (order.user_id, order.created_at))
+            items = self.cursor.fetchone()
+            if items:
+                return jsonify({"msg": "Order successfully added"}), 201
+            return jsonify({"msg": "Order not added"}), 400
+        except psycopg2.IntegrityError:
+            return jsonify({"msg": "Menu Item does not exist"}), 404
+
+    def get_orders(self):
+        """get all orders from database"""
+        return self.query_entire_table('orders')
+
+    def get_single_order(self, order_id):
+        """return a single from the database"""
+        return self.query_single_row('orders', 'order_id', order_id)
+
+    def update_order_status(self, order_id, status):
+        """update the staus of an order"""
+        update_command = "Update orders SET status = '%s' WHERE order_id = '%s'" % (
+            status, order_id)
+        try:
+            self.cursor.execute(
+                "SELECT row_to_json(row) FROM (SELECT * FROM orders WHERE order_id = %s) row;" % (order_id))
+            items = self.cursor.fetchone()
+            if items:
+                self.cursor.execute(update_command)
+                return jsonify({"msg": "Order updated"}), 201
+            else:
+                return jsonify({"msg": "Order not found"}), 404
+        except:
+            return jsonify({"msg": "Error"}), 404
+
+    def get_orders_by_userid(self, user_id):
+        """return all orders belonging to a particular user"""
+        select_command = "SELECT row_to_json(row) FROM (SELECT * FROM orders WHERE user_id = %s) row;" % (
+            user_id)
+        self.cursor.execute(select_command)
+        return self.cursor.fetchall()
